+++ src/PottsScorer.h
@@ -138,7 +138,7 @@ private:
         int hit_count = 0;
     };
 
-    float score_with_sliding_window(const GInterval &original_interval,
+    double score_with_sliding_window(const GInterval &original_interval,
                                     const GInterval &expanded_interval,
                                     size_t i_min, size_t i_max, size_t motif_len,
                                     size_t tlen);
